chore(docs) update README table of contents
In editing the README, I was battling the VSCode markdown plugin's auto-toc functionality, which also revealed a missing section, and what I think is some incorrect indentation in the toc caused by incorrect heading levels. So fi the headings, add some comments to ignore some headings we don't want in the toc, and (auto-)update the toc.